Refactor layout and highlight logic into separate modules
