+++ frontend/flow.js
@@ -12,6 +12,12 @@
   const DedupeUtil = typeof require === 'function'
     ? (() => { try { return require('./src/utils/dedup'); } catch (e) { return {}; } })()
     : (typeof window !== 'undefined' ? (window.Dedupe || {}) : {});
+  const FlowLayout = typeof require === 'function'
+    ? (() => { try { return require('./src/flow/layout'); } catch (e) { return {}; } })()
+    : (typeof window !== 'undefined' ? (window.FlowLayout || {}) : {});
+  const FlowHighlight = typeof require === 'function'
+    ? (() => { try { return require('./src/flow/highlight'); } catch (e) { return {}; } })()
+    : (typeof window !== 'undefined' ? (window.FlowHighlight || {}) : {});
   const parseGedcom = GedcomUtil.parseGedcom || function () { return []; };
   const findBestMatch = DedupeUtil.findBestMatch || function () { return { match: null, score: 0 }; };
   const matchScore = DedupeUtil.matchScore || function () { return 0; };
@@ -107,6 +113,9 @@
             (typeof AppConfig.relativeAttraction !== 'undefined'
               ? parseFloat(AppConfig.relativeAttraction)
               : null)) || 0.5;
+        const LayoutMod = FlowLayout && typeof FlowLayout.createLayoutAPI === 'function' ? FlowLayout.createLayoutAPI({ d3, GenerationLayout, horizontalGridSize, relativeAttraction }) : null;
+        const tidyUpExternal = LayoutMod ? LayoutMod.tidyUp : function() {};
+        const tidyUpChunkedExternal = LayoutMod ? LayoutMod.tidyUpChunked : async function() {};
         loggedIn = ref(window.currentUser && window.currentUser !== 'guest');
         admin = ref(window.isAdmin || false);
         const showDeleteAllButton = computed(
@@ -127,7 +136,7 @@
           setTimeout(() => { flashVisible.value = false; }, 1500);
         }
         function notifyTap() {
-          try { if (navigator.vibrate) navigator.vibrate(10); } catch (e) { /* ignore */ }
+          try { if (typeof navigator !== 'undefined' && navigator.vibrate) navigator.vibrate(10); } catch (e) { /* ignore */ }
         }
         const selected = ref(null);
         const showModal = ref(false);
@@ -197,7 +206,7 @@
         let newNodePos = null;
         let scoreTimer = null;
         // Touch capability detection for mobile-specific UI affordances
-        const isTouchDevice = ref(('ontouchstart' in window) || (navigator.maxTouchPoints > 0));
+        const isTouchDevice = ref((typeof window !== 'undefined' && 'ontouchstart' in window) || (typeof navigator !== 'undefined' && navigator.maxTouchPoints > 0));
         function addClass(edge, cls) {
           const parts = (edge.class || '').split(' ').filter(Boolean);
           if (!parts.includes(cls)) parts.push(cls);
@@ -761,206 +770,32 @@
         // Cache for highlighting state to avoid unnecessary DOM updates
         const highlightedNodes = new Set();
         const highlightedEdges = new Set();
+        let highlight = null;
+        if (FlowHighlight && typeof FlowHighlight.createHighlightAPI === 'function') {
+          highlight = FlowHighlight.createHighlightAPI({
+            getNodes: () => nodes.value,
+            getEdges: () => edges.value,
+            getSelectedEdge: () => selectedEdge.value,
+            getNodeById,
+            childrenCache,
+            addClass,
+            removeClass,
+            highlightedNodes,
+            highlightedEdges,
+          });
+        }
         
         function clearHighlights() {
-          // Only update nodes that were previously highlighted
-          for (const nodeId of highlightedNodes) {
-            const node = nodeMap.get(nodeId);
-            if (node && node.data) node.data.highlight = false;
-          }
-          highlightedNodes.clear();
-          
-          // Only update edges that were previously highlighted
-          for (const edgeId of highlightedEdges) {
-            const edge = edges.value.find(e => e.id === edgeId);
-            if (edge) {
-              // Remove highlight-specific classes but preserve other classes like 'selected-edge'
-              removeClass(edge, 'highlight-edge');
-              removeClass(edge, 'faded-edge');
-              // If no classes remain, remove the class property entirely
-              if (!edge.class || edge.class.trim() === '') {
-                delete edge.class;
-              }
-            }
-          }
-          highlightedEdges.clear();
+          if (highlight) highlight.clearHighlights();
         }
 
         function highlightBloodline(id) {
-          clearHighlights();
-
-          const map = {};
-          nodes.value.forEach((n) => {
-            map[n.id] = n;
-          });
-
-          const visitedUp = new Set();
-          const visitedDown = new Set();
-
-          function unionId(f, m) {
-            return `u-${f}-${m}`;
-          }
-
-          function markNode(nId) {
-            const node = map[String(nId)];
-            if (node && node.data) {
-              node.data.highlight = true;
-              highlightedNodes.add(String(nId));
-            }
-          }
-
-          function highlightAncestors(pid) {
-            if (!pid || visitedUp.has(pid)) return;
-            visitedUp.add(pid);
-            const node = map[String(pid)];
-            if (!node) return;
-            markNode(pid);
-            if (node.data.fatherId && node.data.motherId) {
-              const uId = unionId(node.data.fatherId, node.data.motherId);
-              markNode(uId);
-            }
-            highlightAncestors(node.data.fatherId);
-            highlightAncestors(node.data.motherId);
-          }
-
-          function highlightDescendants(pid) {
-            if (!pid || visitedDown.has(pid)) return;
-            visitedDown.add(pid);
-            const node = map[String(pid)];
-            if (!node) return;
-            markNode(pid);
-            
-            // Use cached children instead of iterating all nodes
-            const children = childrenCache.get(pid) || [];
-            for (const childData of children) {
-              if (childData.fatherId && childData.motherId) {
-                const uId = unionId(childData.fatherId, childData.motherId);
-                markNode(uId);
-              }
-              highlightDescendants(parseInt(childData.id));
-            }
-          }
-
-          highlightAncestors(id);
-          highlightDescendants(id);
-
-          edges.value.forEach((edge) => {
-            const sel = edge === selectedEdge.value;
-            if (edge.id.startsWith('spouse-line')) {
-              edge.class = sel ? 'selected-edge' : 'faded-edge';
-              if (!sel) highlightedEdges.add(edge.id);
-              return;
-            }
-            const src = map[edge.source];
-            const tgt = map[edge.target];
-            if (src?.data.highlight && tgt?.data.highlight) {
-              edge.class = sel ? 'selected-edge' : 'highlight-edge';
-              if (!sel) highlightedEdges.add(edge.id);
-            } else {
-              edge.class = sel ? 'selected-edge' : 'faded-edge';
-              if (!sel) highlightedEdges.add(edge.id);
-            }
-          });
+          if (highlight) highlight.highlightBloodline(id);
         }
 
         // Optimized async version that breaks up work across frames
         async function highlightBloodlineAsync(id) {
-          clearHighlights();
-
-          // Pre-build map for faster lookups
-          const map = new Map();
-          nodes.value.forEach((n) => {
-            map.set(n.id, n);
-          });
-
-          const visitedUp = new Set();
-          const visitedDown = new Set();
-          const toHighlight = new Set();
-          const unionIds = new Set();
-
-          function unionId(f, m) {
-            return `u-${f}-${m}`;
-          }
-
-          // Collect all nodes to highlight first (non-blocking)
-          function collectAncestors(pid) {
-            if (!pid || visitedUp.has(pid)) return;
-            visitedUp.add(pid);
-            const node = map.get(String(pid));
-            if (!node) return;
-            toHighlight.add(pid);
-            if (node.data.fatherId && node.data.motherId) {
-              unionIds.add(unionId(node.data.fatherId, node.data.motherId));
-            }
-            collectAncestors(node.data.fatherId);
-            collectAncestors(node.data.motherId);
-          }
-
-          function collectDescendants(pid) {
-            if (!pid || visitedDown.has(pid)) return;
-            visitedDown.add(pid);
-            const node = map.get(String(pid));
-            if (!node) return;
-            toHighlight.add(pid);
-            
-            // Use cached children instead of iterating all nodes
-            const children = childrenCache.get(pid) || [];
-            for (const childData of children) {
-              if (childData.fatherId && childData.motherId) {
-                unionIds.add(unionId(childData.fatherId, childData.motherId));
-              }
-              collectDescendants(parseInt(childData.id));
-            }
-          }
-
-          // Collect all nodes to highlight
-          collectAncestors(id);
-          collectDescendants(id);
-
-          // Apply highlights in batches to avoid blocking
-          const nodesToUpdate = [...toHighlight, ...unionIds];
-          const batchSize = 50;
-          
-          for (let i = 0; i < nodesToUpdate.length; i += batchSize) {
-            const batch = nodesToUpdate.slice(i, i + batchSize);
-            batch.forEach(nId => {
-              const node = map.get(String(nId));
-              if (node && node.data) node.data.highlight = true;
-            });
-            
-            // Yield control back to browser
-            if (i + batchSize < nodesToUpdate.length) {
-              await new Promise(resolve => setTimeout(resolve, 0));
-            }
-          }
-
-          // Update edges in batches
-          const edgeBatchSize = 100;
-          for (let i = 0; i < edges.value.length; i += edgeBatchSize) {
-            const batch = edges.value.slice(i, i + edgeBatchSize);
-            batch.forEach((edge) => {
-              const sel = edge === selectedEdge.value;
-              if (edge.id.startsWith('spouse-line')) {
-                edge.class = sel ? 'selected-edge' : 'faded-edge';
-                if (!sel) highlightedEdges.add(edge.id);
-                return;
-              }
-              const src = map.get(edge.source);
-              const tgt = map.get(edge.target);
-              if (src?.data.highlight && tgt?.data.highlight) {
-                edge.class = sel ? 'selected-edge' : 'highlight-edge';
-                if (!sel) highlightedEdges.add(edge.id);
-              } else {
-                edge.class = sel ? 'selected-edge' : 'faded-edge';
-                if (!sel) highlightedEdges.add(edge.id);
-              }
-            });
-            
-            // Yield control back to browser
-            if (i + edgeBatchSize < edges.value.length) {
-              await new Promise(resolve => setTimeout(resolve, 0));
-            }
-          }
+          if (highlight) await highlight.highlightBloodlineAsync(id);
         }
 
         function getBloodlineSet(rootId) {
@@ -2346,364 +2181,11 @@
 
         // Chunked version of tidyUp for large datasets
         async function tidyUpChunked(list) {
-          if (list.length === 0) {
-            return;
-          }
-          
-          const CHUNK_SIZE = Math.min(500, Math.max(50, Math.floor(list.length / 10)));
-          
-          const map = new Map(
-            list.map((n) => [n.id, { ...n, children: [], width: n.width || 0 }])
-          );
-          
-          // Process parent-child relationships in chunks
-          let processed = 0;
-          for (const [, node] of map) {
-            if (node.fatherId && map.has(node.fatherId)) {
-              map.get(node.fatherId).children.push(node);
-            } else if (node.motherId && map.has(node.motherId)) {
-              map.get(node.motherId).children.push(node);
-            }
-            
-            processed++;
-            if (processed % CHUNK_SIZE === 0) {
-              await nextTick(); // Yield control to prevent UI freezing
-            }
-          }
-
-          const gen = GenerationLayout.assignGenerations(list);
-          const ROW_HEIGHT = 230;
-
-          const roots = [];
-          map.forEach((n) => {
-            const hasParent = list.some((p) => p.id === n.motherId || p.id === n.fatherId);
-            if (!hasParent) roots.push(n);
-          });
-          
-          const fakeRoot = { id: 'root', children: roots };
-          const baseSpacing = horizontalGridSize * 4;
-          const H_SPACING = baseSpacing - (baseSpacing - horizontalGridSize) * relativeAttraction;
-          const layout = d3.tree().nodeSize([H_SPACING, 1]);
-          const rootNode = d3.hierarchy(fakeRoot);
-          layout(rootNode);
-
-          if (rootNode.children) {
-            rootNode.children.forEach(walk);
-          }
-          function walk(h) {
-            const d = map.get(h.data.id);
-            if (d) {
-              d.x = h.x;
-            }
-            h.children && h.children.forEach(walk);
-          }
-
-          const couples = new Set();
-          // Process couples in chunks
-          for (let i = 0; i < list.length; i += CHUNK_SIZE) {
-            const chunk = list.slice(i, i + CHUNK_SIZE);
-            chunk.forEach((child) => {
-              if (child.fatherId && child.motherId) {
-                const key = `${child.fatherId}-${child.motherId}`;
-                if (!couples.has(key)) {
-                  couples.add(key);
-                  const father = map.get(child.fatherId);
-                  const mother = map.get(child.motherId);
-                  if (father && mother) {
-                    const mid =
-                      (father.x + father.width / 2 + mother.x + mother.width / 2) /
-                      2;
-                    father.x = mid - father.width / 2 - H_SPACING / 2;
-                    mother.x = mid + H_SPACING / 2 - mother.width / 2;
-                  }
-                }
-              }
-            });
-            
-            if (i + CHUNK_SIZE < list.length) {
-              await nextTick(); // Yield control
-            }
-          }
-
-          const links = [];
-          // Process links in chunks
-          for (let i = 0; i < list.length; i += CHUNK_SIZE) {
-            const chunk = list.slice(i, i + CHUNK_SIZE);
-            chunk.forEach((p) => {
-              if (p.fatherId && map.has(p.fatherId)) {
-                links.push({ source: map.get(p.id), target: map.get(p.fatherId), type: 'parent' });
-              }
-              if (p.motherId && map.has(p.motherId)) {
-                links.push({ source: map.get(p.id), target: map.get(p.motherId), type: 'parent' });
-              }
-              (p.spouseIds || []).forEach((sid) => {
-                if (map.has(sid)) {
-                  links.push({ source: map.get(p.id), target: map.get(sid), type: 'spouse' });
-                }
-              });
-            });
-            
-            if (i + CHUNK_SIZE < list.length) {
-              await nextTick(); // Yield control
-            }
-          }
-
-          const nodesForSim = Array.from(map.values());
-          nodesForSim.forEach((n) => {
-            const g = gen.get(n.id) ?? 0;
-            n.y = g * ROW_HEIGHT;
-            n.fy = n.y;
-          });
-
-          const linkForce = d3
-            .forceLink(links)
-            .id((d) => d.id)
-            .distance((d) => (d.type === 'spouse' ? H_SPACING / 2 : 0))
-            .strength(1);
-          const collideForce = d3
-            .forceCollide()
-            .radius((d) => (d.width || 0) / 2 + H_SPACING / 2)
-            .strength(1);
-          const sim = d3
-            .forceSimulation(nodesForSim)
-            .force('link', linkForce)
-            .force('collide', collideForce)
-            .alphaDecay(0.05)
-            .velocityDecay(0.4)
-            .stop();
-          
-          // Run simulation with periodic yielding for large datasets
-          const SIMULATION_CHUNK = 10;
-          for (let i = 0; i < 120; i += SIMULATION_CHUNK) {
-            const iterations = Math.min(SIMULATION_CHUNK, 120 - i);
-            for (let j = 0; j < iterations; j++) {
-              sim.tick();
-            }
-            if (i + SIMULATION_CHUNK < 120) {
-              await nextTick(); // Yield control periodically during simulation
-            }
-          }
-          
-          nodesForSim.forEach((n) => {
-            delete n.fy;
-          });
-
-          // Continue with the rest of the original tidyUp logic...
-          const rows = new Map();
-          nodesForSim.forEach((n) => {
-            const row = Math.round(n.y / ROW_HEIGHT);
-            if (!rows.has(row)) rows.set(row, []);
-            rows.get(row).push(n);
-          });
-
-          rows.forEach((rowNodes) => {
-            rowNodes.sort((a, b) => a.x - b.x);
-            let lastX = Number.NEGATIVE_INFINITY;
-            rowNodes.forEach((n) => {
-              if (n.x < lastX + (n.width || 0) / 2 + H_SPACING / 2) {
-                n.x = lastX + (n.width || 0) / 2 + H_SPACING / 2;
-              }
-              lastX = n.x + (n.width || 0) / 2;
-            });
-          });
-
-          list.forEach((original) => {
-            const updated = map.get(original.id);
-            if (updated) {
-              original.x = updated.x;
-              original.y = updated.y;
-            }
-          });
+          return tidyUpChunkedExternal(list);
         }
 
         function tidyUp(list) {
-          if (list.length === 0) {
-            return;
-          }
-          
-          const map = new Map(
-            list.map((n) => [n.id, { ...n, children: [], width: n.width || 0 }])
-          );
-          map.forEach((n) => {
-            if (n.fatherId && map.has(n.fatherId)) {
-              map.get(n.fatherId).children.push(n);
-            } else if (n.motherId && map.has(n.motherId)) {
-              map.get(n.motherId).children.push(n);
-            }
-          });
-
-          const gen = GenerationLayout.assignGenerations(list);
-          const ROW_HEIGHT = 230;
-
-          const roots = [];
-          map.forEach((n) => {
-            const hasParent = list.some((p) => p.id === n.motherId || p.id === n.fatherId);
-            if (!hasParent) roots.push(n);
-          });
-          const fakeRoot = { id: 'root', children: roots };
-          const baseSpacing = horizontalGridSize * 4;
-          const H_SPACING = baseSpacing - (baseSpacing - horizontalGridSize) * relativeAttraction;
-          const layout = d3.tree().nodeSize([H_SPACING, 1]);
-          const rootNode = d3.hierarchy(fakeRoot);
-          layout(rootNode);
-
-          if (rootNode.children) {
-            rootNode.children.forEach(walk);
-          }
-          function walk(h) {
-            const d = map.get(h.data.id);
-            if (d) {
-              d.x = h.x;
-            }
-            h.children && h.children.forEach(walk);
-          }
-
-
-          const couples = new Set();
-          list.forEach((child) => {
-            if (child.fatherId && child.motherId) {
-              const key = `${child.fatherId}-${child.motherId}`;
-              if (!couples.has(key)) {
-                couples.add(key);
-                const father = map.get(child.fatherId);
-                const mother = map.get(child.motherId);
-                if (father && mother) {
-                  const mid =
-                    (father.x + father.width / 2 + mother.x + mother.width / 2) /
-                    2;
-                  father.x = mid - father.width / 2 - H_SPACING / 2;
-                  mother.x = mid + H_SPACING / 2 - mother.width / 2;
-                }
-              }
-            }
-          });
-
-          const links = [];
-          list.forEach((p) => {
-            if (p.fatherId && map.has(p.fatherId)) {
-              links.push({ source: map.get(p.id), target: map.get(p.fatherId), type: 'parent' });
-            }
-            if (p.motherId && map.has(p.motherId)) {
-              links.push({ source: map.get(p.id), target: map.get(p.motherId), type: 'parent' });
-            }
-            (p.spouseIds || []).forEach((sid) => {
-              if (map.has(sid)) {
-                links.push({ source: map.get(p.id), target: map.get(sid), type: 'spouse' });
-              }
-            });
-          });
-
-          const nodesForSim = Array.from(map.values());
-          nodesForSim.forEach((n) => {
-            const g = gen.get(n.id) ?? 0;
-            n.y = g * ROW_HEIGHT;
-            n.fy = n.y;
-          });
-
-          const linkForce = d3
-            .forceLink(links)
-            .id((d) => d.id)
-            .distance((d) => (d.type === 'spouse' ? H_SPACING / 2 : 0))
-            .strength(1);
-          const collideForce = d3
-            .forceCollide()
-            .radius((d) => (d.width || 0) / 2 + H_SPACING / 2)
-            .strength(1);
-          const sim = d3
-            .forceSimulation(nodesForSim)
-            .force('link', linkForce)
-            .force('collide', collideForce)
-            .alphaDecay(0.05)
-            .velocityDecay(0.4)
-            .stop();
-          for (let i = 0; i < 120; i++) sim.tick();
-          nodesForSim.forEach((n) => {
-            delete n.fy;
-          });
-
-          const rows = new Map();
-          list.forEach((n) => {
-            const g = gen.get(n.id) ?? 0;
-            if (!rows.has(g)) rows.set(g, []);
-            rows.get(g).push(map.get(n.id));
-          });
-
-          rows.forEach((row) => {
-            row.forEach((p) => {
-              if (typeof p.x !== 'number') p.x = 0;
-            });
-            row.sort((a, b) => a.x - b.x);
-            for (let i = 1; i < row.length; i++) {
-              const prev = row[i - 1];
-              const curr = row[i];
-              const prevRight = prev.x + (prev.width || 0);
-              const minX = prevRight + H_SPACING;
-              if (curr.x < minX) {
-                curr.x = minX;
-              }
-            }
-
-            const idMap = new Map(row.map((n) => [n.id, n]));
-            const parentUF = {};
-            const find = (id) => {
-              if (!(id in parentUF)) parentUF[id] = id;
-              if (parentUF[id] !== id) parentUF[id] = find(parentUF[id]);
-              return parentUF[id];
-            };
-            const unite = (a, b) => {
-              const pa = find(a);
-              const pb = find(b);
-              if (pa !== pb) parentUF[pb] = pa;
-            };
-            row.forEach((n) => {
-              (n.spouseIds || []).forEach((sid) => {
-                if (idMap.has(sid)) unite(n.id, sid);
-              });
-            });
-            const groupsMap = {};
-            row.forEach((n) => {
-              const root = find(n.id);
-              groupsMap[root] = groupsMap[root] || [];
-              groupsMap[root].push(n);
-            });
-            const groups = Object.values(groupsMap);
-            groups.forEach((g) => {
-              g.sort((a, b) => a.x - b.x);
-              for (let i = 1; i < g.length; i++) {
-                const prev = g[i - 1];
-                const curr = g[i];
-                const minX = prev.x + (prev.width || 0) + H_SPACING;
-                if (curr.x < minX) {
-                  const shift = minX - curr.x;
-                  for (let j = i; j < g.length; j++) {
-                    g[j].x += shift;
-                  }
-                }
-              }
-            });
-            groups.sort((a, b) => a[0].x - b[0].x);
-            for (let i = 1; i < groups.length; i++) {
-              const prevG = groups[i - 1];
-              const currG = groups[i];
-              const prevRight =
-                prevG[prevG.length - 1].x +
-                (prevG[prevG.length - 1].width || 0);
-              const minX = prevRight + H_SPACING;
-              if (currG[0].x < minX) {
-                const shift = minX - currG[0].x;
-                currG.forEach((n) => {
-                  n.x += shift;
-                });
-              }
-            }
-          });
-
-          list.forEach((n) => {
-            const p = map.get(n.id);
-            n.x = p.x;
-            const g = gen.get(n.id) ?? 0;
-            n.y = g * ROW_HEIGHT;
-          });
+          return tidyUpExternal(list);
         }
