+++ frontend/jest.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-  testEnvironment: 'node'
+  testEnvironment: 'jsdom'
 };
